+++ src/main/java/es/in2/issuer/backend/issuance/application/workflow/IssuanceWorkflow.java
@@ -7,9 +7,9 @@ import reactor.core.publisher.Mono;
 public interface IssuanceWorkflow {
 
     Mono<IssuanceResponse> issueCredential(String processId, IssuanceRequest request, String idToken,
-                                           String publicIssuerBaseUrl);
+                                           String publicIssuerBaseUrl, String publicWalletBaseUrl);
 
     Mono<IssuanceResponse> issueCredentialWithoutAuthorization(String processId, IssuanceRequest request, String token,
-                                                               String publicIssuerBaseUrl);
+                                                               String publicIssuerBaseUrl, String publicWalletBaseUrl);
 
 }